# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: SABABUKA Bersinar API
  version: 1.0.0-draft
  summary: Kontrak API MVP untuk tata kelola data Kabupaten Kapuas
  description: |
    API tunggal untuk ruang Superadmin, BAPPERIDA, Kominfo, OPD, Pimpinan,
    dashboard eksekutif, dan Asisten AI. Semua operasi dibatasi oleh permission,
    lingkup organisasi, status workflow, dan feature flag.
  contact:
    name: Pengelola SABABUKA Bersinar
  license:
    name: Penggunaan Internal Pemerintah Kabupaten Kapuas
    identifier: LicenseRef-SABABUKA-Internal
servers:
  - url: http://127.0.0.1:3001/api/v1
    description: Local development (aktif saat ini)
  - url: https://dev.sababuka.com/api/v1
    description: Development (belum diaktifkan)
  - url: https://sababuka.com/api/v1
    description: Production (setelah disetujui)
...
security:
  - sessionCookie: []
paths:
...
  /auth/login:
    post:
      tags: [Authentication]
      operationId: login
      summary: Membuat sesi pengguna
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login berhasil; cookie sesi dikirim melalui Set-Cookie
          headers:
            Set-Cookie:
              schema: {type: string}
              description: Cookie sesi HttpOnly, Secure, SameSite=Lax
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
  /auth/logout:
    post:
      tags: [Authentication]
      operationId: logout
      summary: Mengakhiri sesi aktif
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      responses:
        "204": {description: Sesi diakhiri}
        "401": {$ref: "#/components/responses/Unauthorized"}
  /auth/invitations/{token}:
    get:
      tags: [Authentication]
      operationId: inspectInvitation
      summary: Memeriksa undangan pengguna yang masih berlaku
      security: []
      parameters:
        - $ref: "#/components/parameters/InvitationToken"
      responses:
        "200":
          description: Informasi minimum undangan
          content:
            application/json:
              schema: {$ref: "#/components/schemas/InvitationInfo"}
        "404": {$ref: "#/components/responses/NotFound"}
  /auth/invitations/{token}/mfa/setup:
    post:
      tags: [Authentication]
      operationId: setupInvitationMfa
      summary: Membuat secret TOTP untuk pengguna undangan
      security: []
      parameters:
        - $ref: "#/components/parameters/InvitationToken"
      responses:
        "200":
          description: Secret dan URI provisioning Authenticator; hanya ditampilkan pada tahap ini
          content:
            application/json:
              schema: {$ref: "#/components/schemas/TotpSetupResponse"}
        "404": {$ref: "#/components/responses/NotFound"}
        "429": {$ref: "#/components/responses/RateLimited"}
  /auth/invitations/{token}/accept:
    post:
      tags: [Authentication]
      operationId: acceptInvitation
      summary: Mengaktifkan akun dengan password dan verifikasi TOTP
      security: []
      parameters:
        - $ref: "#/components/parameters/InvitationToken"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/InvitationAcceptInput"}
      responses:
        "200":
          description: Akun aktif; recovery code hanya dikirim sekali
          content:
            application/json:
              schema: {$ref: "#/components/schemas/InvitationAcceptResponse"}
        "400": {$ref: "#/components/responses/ValidationError"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
        "429": {$ref: "#/components/responses/RateLimited"}
...
  /me/mfa/totp/setup:
    post:
      tags: [Authentication]
      operationId: setupCurrentUserMfa
      summary: Memulai enrollment TOTP untuk pengguna aktif
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      responses:
        "200":
          description: Secret dan URI provisioning Authenticator
          content:
            application/json:
              schema: {$ref: "#/components/schemas/TotpSetupResponse"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "409": {$ref: "#/components/responses/Conflict"}
  /me/mfa/totp/confirm:
    post:
      tags: [Authentication]
      operationId: confirmCurrentUserMfa
      summary: Memverifikasi enrollment TOTP dan menghasilkan recovery code
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/MfaConfirmInput"}
      responses:
        "200":
          description: MFA aktif; recovery code hanya dikirim sekali
          content:
            application/json:
              schema: {$ref: "#/components/schemas/RecoveryCodesResponse"}
        "400": {$ref: "#/components/responses/ValidationError"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "404": {$ref: "#/components/responses/NotFound"}
...
  /assistant/sessions:
    post:
      tags: [Assistant]
      operationId: createAssistantSession
      summary: Membuat sesi Asisten AI
      description: Membutuhkan `assistant.use` dan feature flag `assistant.enabled`.
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title: {type: string, maxLength: 255}
      responses:
        "201":
          description: Sesi dibuat
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AssistantSession"}
        "403": {$ref: "#/components/responses/Forbidden"}
  /assistant/sessions/{session_id}/messages:
    post:
      tags: [Assistant]
      operationId: sendAssistantMessage
      summary: Mengajukan pertanyaan kepada Asisten AI
      description: |
        Jawaban hanya menggunakan publikasi aktif yang dapat diakses pengguna.
        Jika data tidak cukup, respons menyatakan keterbatasan dan tidak mengarang angka.
      parameters:
        - name: session_id
          in: path
          required: true
          schema: {$ref: "#/components/schemas/Uuid"}
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message: {type: string, minLength: 1, maxLength: 4000}
      responses:
        "200":
          description: Jawaban beserta sitasi data
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AssistantAnswer"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "422": {$ref: "#/components/responses/BusinessRuleViolation"}
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: sababuka_session
      description: Sesi server-side; cookie wajib HttpOnly, Secure, dan SameSite=Lax.
  parameters:
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema: {type: string, minLength: 20}
      description: Token CSRF terikat ke sesi untuk seluruh operasi mutasi.
...
    InvitationToken:
      name: token
      in: path
      required: true
      schema: {type: string, minLength: 32, maxLength: 200}
      description: Token acak sekali pakai; nilai asli tidak disimpan di database.
...
  responses:
    Unauthorized:
      description: Sesi tidak tersedia atau kedaluwarsa
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ApiError"}
          example:
            error: {code: AUTH_REQUIRED, message: Sesi diperlukan., request_id: 8be45d40-e437-4f44-86da-91d0cae75488}
    Forbidden:
      description: Permission atau scope tidak mencukupi
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ApiError"}
    NotFound:
      description: Objek tidak ditemukan atau tidak boleh diungkapkan
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ApiError"}
    ValidationError:
      description: Permintaan tidak valid
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ApiError"}
...
    Conflict:
      description: Konflik kode, versi, atau data aktif
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ApiError"}
    BusinessRuleViolation:
      description: Aturan bisnis tidak terpenuhi
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ApiError"}
...
    RateLimited:
      description: Terlalu banyak permintaan
      headers:
        Retry-After:
          schema: {type: integer}
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ApiError"}
...
  schemas:
    Uuid:
      type: string
      format: uuid
    Timestamp:
      type: string
      format: date-time
...
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code: {type: string}
            message: {type: string}
            request_id: {$ref: "#/components/schemas/Uuid"}
            details:
              type: array
              items:
                type: object
                additionalProperties: true
...
    LoginRequest:
      type: object
      additionalProperties: false
      description: "mfa_code dan recovery_code saling eksklusif; kirim salah satu ketika akun mewajibkan MFA."
      required: [identifier, password]
      properties:
        identifier: {type: string, minLength: 3, maxLength: 255}
        password: {type: string, minLength: 8, maxLength: 256, writeOnly: true}
        mfa_code: {type: string, pattern: "^[0-9]{6}$", writeOnly: true}
        recovery_code: {type: string, pattern: "^[A-Za-z0-9-]{16,32}$", writeOnly: true}
    SessionResponse:
      type: object
      required: [user, csrf_token, expires_at]
      properties:
        user: {$ref: "#/components/schemas/CurrentUser"}
        csrf_token: {type: string, writeOnly: true}
        expires_at: {$ref: "#/components/schemas/Timestamp"}
    InvitationInfo:
      type: object
      required: [email, full_name, expires_at, mfa_setup_started]
      properties:
        email: {type: string, format: email}
        full_name: {type: string}
        expires_at: {$ref: "#/components/schemas/Timestamp"}
        mfa_setup_started: {type: boolean}
    TotpSetupResponse:
      type: object
      required: [secret, provisioning_uri, issuer, account]
      properties:
        secret: {type: string, pattern: "^[A-Z2-7]+$", description: Nilai sensitif yang hanya ditampilkan saat enrollment.}
        provisioning_uri: {type: string, pattern: "^otpauth://totp/"}
        issuer: {type: string}
        account: {type: string}
    InvitationAcceptInput:
      type: object
      additionalProperties: false
      required: [password, mfa_code]
      properties:
        password: {type: string, minLength: 12, maxLength: 256, writeOnly: true}
        mfa_code: {type: string, pattern: "^[0-9]{6}$", writeOnly: true}
    MfaConfirmInput:
      type: object
      additionalProperties: false
      required: [mfa_code]
      properties:
        mfa_code: {type: string, pattern: "^[0-9]{6}$", writeOnly: true}
    RecoveryCodesResponse:
      type: object
      required: [recovery_codes]
      properties:
        recovery_codes:
          type: array
          minItems: 8
          maxItems: 8
          uniqueItems: true
          items: {type: string, pattern: "^[A-F0-9]{4}(?:-[A-F0-9]{4}){3}$"}
    InvitationAcceptResponse:
      allOf:
        - $ref: "#/components/schemas/RecoveryCodesResponse"
        - type: object
          required: [email, status]
          properties:
            email: {type: string, format: email}
            status: {type: string, const: active}
    CurrentUser:
      type: object
      required: [id, email, full_name, mfa_required, roles, permissions, organizations]
      properties:
        id: {$ref: "#/components/schemas/Uuid"}
        email: {type: string, format: email}
        full_name: {type: string}
        mfa_required: {type: boolean}
        roles:
          type: array
          items: {$ref: "#/components/schemas/EffectiveRole"}
        permissions:
          type: array
          uniqueItems: true
          items: {type: string}
        organizations:
          type: array
          items: {$ref: "#/components/schemas/Organization"}
    EffectiveRole:
      type: object
      required: [code, scope_type]
      properties:
        code: {type: string}
        scope_type: {type: string, enum: [global, organization, self, published]}
        organization_id:
          oneOf:
            - $ref: "#/components/schemas/Uuid"
            - type: "null"
...
    OrganizationInput:
      type: object
      additionalProperties: false
      required: [code, name, organization_type]
      properties:
        code: {type: string, pattern: "^[A-Z0-9][A-Z0-9._-]*$", maxLength: 64}
        name: {type: string, maxLength: 255}
        short_name: {type: [string, "null"], maxLength: 120}
        organization_type: {type: string, maxLength: 40}
        parent_id:
          oneOf:
            - $ref: "#/components/schemas/Uuid"
            - type: "null"
    Organization:
      allOf:
        - $ref: "#/components/schemas/OrganizationInput"
        - type: object
          required: [id, is_active, created_at, updated_at]
          properties:
            id: {$ref: "#/components/schemas/Uuid"}
            is_active: {type: boolean}
            created_at: {$ref: "#/components/schemas/Timestamp"}
            updated_at: {$ref: "#/components/schemas/Timestamp"}
...
    AssistantSession:
      type: object
      required: [id, status, started_at]
      properties:
        id: {$ref: "#/components/schemas/Uuid"}
        title: {type: [string, "null"]}
        status: {type: string, enum: [active, closed, archived]}
        started_at: {$ref: "#/components/schemas/Timestamp"}
    AssistantAnswer:
      type: object
      required: [message_id, answer, citations, sufficiency]
      properties:
        message_id: {$ref: "#/components/schemas/Uuid"}
        answer: {type: string}
        sufficiency: {type: string, enum: [sufficient, partial, insufficient]}
        citations:
          type: array
          items:
            type: object
            required: [publication_item_id, label]
            properties:
              publication_item_id: {$ref: "#/components/schemas/Uuid"}
              label: {type: string}
              indicator: {type: string}
              period: {type: string}
              organization: {type: string}
              source: {type: string}
...